requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0
kafka-python==2.0.2
pyarrow==14.0.1
delta-spark==3.0.0
mlflow==2.9.2
great-expectations==0.18.2
